+++ ModuleParticles.h
@@ -93,8 +93,8 @@ public:
 	bool cleanUp();
 	void onCollision(Collider *col1, Collider *col2);
 
-	void addWeapon(WEAPON_TYPES type , int x, int y, COLLIDER_TYPE collider_type = COLLIDER_NONE, Uint32 delay = 0);
-	void addExplosion(EXPLOSION_TYPES type, int x, int y, Uint32 delay = 0);
+	void addWeapon(WEAPON_TYPES type , float x, float y, COLLIDER_TYPE collider_type = COLLIDER_NONE, Uint32 delay = 0);
+	void addExplosion(EXPLOSION_TYPES type, float x, float y, Uint32 delay = 0);
 };
 
 #endif //!__MODULEPARTICLES_H__
